feat: 0.0.2 WIP rollup
core:
- Artist / scene_index: thin-instance bond buffer refactor (bond_buffer.ts)
- Manipulate mode rework
- Draw command + reader tweaks
- Marching cubes test + WASM smoke test

page:
- LeftSidebar and ClusterPanel redesign
- ScreenshotDialog polish
- MolvisWrapper / BuilderTab / DownloadStructureSection tweaks
- Tailwind styles update
- rstest config + wasm serialization test

docs:
- molrs-0.0.8 precision audit

Root:
- Rename 'build' → 'build:all' so npm run build aliases don't shadow
  workspace-level build scripts.

+++ core/examples/demo_frame.ts
@@ -31,7 +31,7 @@ async function main() {
   bondsBlock.setColU32("j", new Uint32Array([1, 2]));
   bondsBlock.setColU32("order", new Uint32Array([1, 1]));
 
-  const box = Box.cube(1, new Float32Array([0, 0, 0]), false, false, false);
+  const box = Box.cube(1, new Float64Array([0, 0, 0]), false, false, false);
 
   const frame = new Frame();
   frame.insertBlock("atoms", atomsBlock);
